+++ src/observer/time.hxx
@@ -28,7 +28,7 @@ public:
 	TimeObserver();
 	virtual ~TimeObserver() {};
 
-	virtual void setFpb(int fpb) {};
+	virtual void setFpb(double fpb) {};
 
 	//Reset any internal set of this object regarding time. Don't reset/delete any buffers!!
 	//This is not to be confused with Stately::reset()
